Commiting 1.03 from CPAN
This repo is so old and I have not worked on it in so long that I no longer know
where I left the repository. That was 5 laptops or so ago. I'm sure it's in a
backup drive somewhere... but it'll be in Subversion and whatever... let's just
start over.